fixes #19
helps to save constructor parameters to respective variables
