+++ nbproject/project.properties
@@ -82,7 +82,7 @@ jnlp.signed=false
 jnlp.signing=
 jnlp.signing.alias=
 jnlp.signing.keystore=
-main.class=com.shsrobotics.reinforcementlearning.util.Test
+main.class=com.shsrobotics.reinforcementlearning.estimators.Test
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
 platform.active=default_platform
